src/goStatus.ts: refactor status bar code into goStatus.ts
Refactor the status bar environment and entry items into the same
goStatus.ts file.

Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/258520
Trust: Suzy Mueller <suzmue@golang.org>
Trust: Hyang-Ah Hana Kim <hyangah@gmail.com>
Run-TryBot: Suzy Mueller <suzmue@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
