+++ src/goMode.ts
@@ -8,3 +8,14 @@
 import vscode = require('vscode');
 
 export const GO_MODE: vscode.DocumentFilter = { language: 'go', scheme: 'file' };
+export const GO_MOD_MODE: vscode.DocumentFilter = { language: 'go.mod', scheme: 'file' };
+export const GO_SUM_MODE: vscode.DocumentFilter = { language: 'go.sum', scheme: 'file' };
+
+export function isGoFile(document: vscode.TextDocument): boolean {
+	if (vscode.languages.match(GO_MODE, document)
+	|| vscode.languages.match(GO_MOD_MODE, document)
+	|| vscode.languages.match(GO_SUM_MODE, document)) {
+		return true;
+	}
+	return false;
+}
